requirements.txt
selenium==4.15.0
pandas==2.1.0
selenium-stealth==1.0.6
pika==1.3.0
python-dotenv==1.0.0